+++ examples/localclient_linux/doc.go
@@ -1,0 +1,3 @@
+// Example localclient_linux is a demonstration of local Linux and VPP config
+// management via clientv1.
+package main