+++ plugins/defaultplugins/aclplugin/acl_config.go
@@ -17,7 +17,8 @@
 //go:generate binapi-generator --input-file=/usr/share/vpp/api/acl.api.json --output-dir=bin_api
 
 
-// Package aclplugin implements the ACL Plugin.
+// Package aclplugin implements the ACL Plugin that handles management of VPP
+// Access lists.
 package aclplugin
 
 import (