+++ plugins/defaultplugins/doc.go
@@ -1,0 +1,3 @@
+// Package defaultplugins is the parent package for all "default" VPP-related
+// plugins.
+package defaultplugins