+++ plugins/defaultplugins/l2plugin/vppdump/doc.go
@@ -1,0 +1,3 @@
+// Package vppdump provides helpers to dump all bridge-domains, L2 FIBs and
+// XConnect pairs configured in VPP.
+package vppdump