+++ plugins/defaultplugins/l3plugin/vppdump/doc.go
@@ -1,0 +1,2 @@
+// Package vppdump provides helpers to dump all L3 FIBs configured in VPP.
+package vppdump