+++ scripts/build-vpp.sh
@@ -2,7 +2,8 @@
 set -e
 
 VPP_CACHE_DIR=$HOME/build-cache/vpp
-VPP_COMMIT="8e4222fc7e23a478b021930ade3cb7d20938e398"
+VPP_COMMIT="9969a3274bfdbaa62b3528887641c5ea36a8509f"
+VPP_IMG_TAG="9969a32"
 
 if [ ! -d "$VPP_CACHE_DIR" ]; then
     echo "Building VPP binaries."