+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# Release v1.0.5 (2017-9-25)
+
+## Compatibility
+VPP version v17.10-rc0~370-g9969a32
+
+## Major Themes
+
+- [GoVppMux](plugins/govppmux)
+    - configuration file for govpp added
+- [Kafka Partitions](messaging/kafka)
+    - Changes in offset handling, only automatically partitioned messages (hash, random)
+      have their offset marked. Manually partitioned messages are not marked.
+    - Implemented post-init consumer (for manual partitioner only) which allows to start
+      consuming after kafka-plugin Init()   
+    - Minimalistic examples & documentation for Kafka API will be improved in a later release.
+    
 # Release v1.0.4 (2017-09-08)
 
 ## Major Themes